refactor(transport): clarify package boundaries
The Transport interface only lives in 'internal'
because it's shared by public clients. The 'internal' package,
however, does not have an authority on what the implementation
should be. Meaning it cannot have a NewTransport function.
All 'internal' package does is define _interfaces_ for
Transport, Request, Endpoint, and Message.

The productive implementation lives in the 'internal/transport'
package. It defines its own transport.Config and exports 1 public
struct, transport.T, and a transport.New function to construct it.

The public weaviate.Client calls transport.New directly, instead
of relying on the 'internal' package to provide an implementation.