refactor(transport): remove dependency on 'api' package
We further deliniate the boundaries between the 'internal',
''transport', and 'api' packages.

Firstly, Endpoint and Message interfaces are now declared
in the 'transport' package, where they are used. The logic is simple:
'transport' package provides a Transport implementation that supports
_REST and gRPC_ requests. This gives us an even leaner Transport iface,
as it now does not hint of possible implementations.

Request is anything that can have a body.

Secondly, we cut the dependency on gRPC-related 'api' structs in the
'transport' package. The only thing necessary to send a gRPC request
is a valid protobuf stub. The transport also doesn't care how the
reply stub is unmarshaled, so 'dest' stays opaque here.
We encode that in the Message[Request, Reply] interafce, which the
structs in the 'api' package now implement.
Compile-time assertions ensure these interfaces are implemented
correclty.