chore: update generated models and stubs
In weaviate/weaviate protobuf stubs are generated in the 'protocol'
package. Rather than tinker with the .proto files, we will just alias
protocol => proto wherever we use it. Luckily it's not too many places.

The OpenAPI schema has some new fields so the models have been regenerated.
Some are deprecated, but this commit does not remove their usage in the client (yet).

CleanupIntervalSeconds is changed to int32, because it's type in Swagger v2 spec
was 'int' and not 'int64' or 'uint64'. This logic is encoded in ./mage/magefile.go.