refactor(query): unify vector target handling for embeddings and text
With the new design, users can use the same syntax and functions
for defining multi-vector targets both in vector- and text-based
searches.
Rather than requiring a bulky NearVectorTarget interface in internal/api,
we present a concrete SearchTarget struct that has all fields necessary
to build a request.
Instead, we push the VectorKind and VectorTarget interfaces up into the
public part of the API, where the distinction between different input types exits:
users may pass a vector embedding, or a text string, assign a weight or not.
By defining query.VectorName type we are able to support all use-cases
with the same set of aggregation functions: Sum, Min, Average, ManualWeights, and RelativeScore.

As a result, we're able to get rid of the now-redundant copies of
api.NearVectorTarget and api.TargetVector implementations in internal/message_test.go